requirements/base.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.94.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
jose==1.0.0
passlib==1.7.4
pydantic==1.10.6
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
starlette==0.26.0.post1
typing_extensions==4.5.0
uvicorn==0.21.0
watchfiles==0.18.1
websockets==10.4

# Celery
celery==5.2.7
flower==1.2.0

# Database

# Postgres
psycopg2==2.9.5
